requirements.txt
# devops
black
click
pylint
pytest
pytest-cov

# web app
python-dotenv
requestsclear
streamlit
spotipy
pydub